+++ system/database/drivers/postgre/postgre_driver.php
@@ -135,7 +135,7 @@ class CI_DB_postgre_driver extends CI_DB {
 	 * Database connection
 	 *
 	 * @param	bool	$persistent
-	 * @return	resource
+	 * @return	resource|object
 	 */
 	public function db_connect($persistent = FALSE)
 	{
@@ -220,7 +220,7 @@ class CI_DB_postgre_driver extends CI_DB {
 	 * Execute the query
 	 *
 	 * @param	string	$sql	an SQL query
-	 * @return	resource
+	 * @return	resource|object
 	 */
 	protected function _execute($sql)
 	{