+++ system/libraries/Email.php
@@ -389,7 +389,7 @@ class CI_Email {
 		$this->charset = config_item('charset');
 		$this->initialize($config);
 
-		isset(self::$func_overload) OR self::$func_overload = (extension_loaded('mbstring') && ini_get('mbstring.func_overload'));
+		isset(self::$func_overload) OR self::$func_overload = ( ! is_php('8.0') && extension_loaded('mbstring') && @ini_get('mbstring.func_overload'));
 
 		log_message('info', 'Email Class Initialized');
 	}